+++ packages/cli/src/rules/index.ts
@@ -4,10 +4,12 @@ import { ccipRules } from "./ccip.js";
 import { dataFeedRules } from "./data-feeds.js";
 import { dataStreamsRules } from "./data-streams.js";
 import { functionsCreRules } from "./functions-cre.js";
+import { registryRules } from "./registry-checks.js";
 import { vrfRules } from "./vrf.js";
 
 export const rules: Rule[] = [
   ...dataFeedRules,
+  ...registryRules,
   ...dataStreamsRules,
   ...ccipRules,
   ...vrfRules,